Rename user change endpoint classes
